+++ picture_editor/gaussblur.h
@@ -1,6 +1,7 @@
 #ifndef GAUSSBLUR_H
 #define GAUSSBLUR_H
 
+#include <iostream>
 #include <QVector>
 #include <opencv2/core/core.hpp>
 #include <opencv2/highgui/highgui.hpp>
@@ -11,7 +12,7 @@ class GaussBlur
 {
 public:
     GaussBlur();
-    float calculateKernel(int kernel, cv::Mat img, int channel, int i, int j);
+    int calculateKernel(int kernel, cv::Mat img, int channel, int i, int j);
     void calculatePixel(cv::Mat img, cv::Mat helper, int i, int j, int kernel);
 
 private:
